rustls_pki_types

Module server_name

DNS name validation according to RFC1035, but with underscores allowed.

Modules

parser 🔒

Structs

AddrParseError

Failure to parse an IP address

DnsName

A type which encapsulates a string (borrowed or owned) that is a syntactically valid DNS name.

InvalidDnsNameError

The provided input could not be parsed because it is not a syntactically-valid DNS Name.

Ipv4Addr

no_std implementation of std::net::Ipv4Addr.

Ipv6Addr

no_std implementation of std::net::Ipv6Addr.

Enums

DnsNameInner 🔒

IpAddr

no_std implementation of std::net::IpAddr.

ServerName

Encodes ways a client can know the expected name of the server.

Functions

validate 🔒